adjust working, need to move updating product inventory from shipment recieve to shipment approve

+++ database/migrations/2025_09_04_041351_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name')->unique();
-            $table->bigInteger('quantity')->default(0);
+            $table->float('quantity', 10, 2)->default(0);
             $table->foreignId('category_id')->constrained();    
             // $table->foreignId('vendor_id')->constrained();  
             $table->text('description')->nullable();
